Add files via upload
This Program used to find the Duplicate characters in the string 
Example:
Here, you can see we have passed "google" as a string, and we got a result as "og". It means these two characters (og)are duplicates in the word "google".

and to find the Unique characters in the string

Example:
you can see the output as"gole". This means after removing duplicate characters from the word "google," we get the "gole" which are unique chars.